Fix last warnings
